02_langchain/pyproject.toml
[project]
name = "02-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "altair<5",
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "google-genai>=1.63.0",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.19.0",
    "tqdm>=4.67.3",
]
